// src/app/models/Post.ts
export interface Post {
  category: string;
  title: string;
  photoMalumot: string;
  textMalumot: string;
  photoTarikh: string;
  textTarikh: string;
  photoTajhizot: string;
  textTajhizot: string;
  photoTamos: string;
  textTamos: string;
  photoLocation: string;
  textLocation: string;
  photoFaoliyat: string;
  textFaoliyat: string;
  photoAsarho: string;
  textAsarho: string;
  date: string;
  author: string;
  _id: string;
}